Switch to a background work-queue model
This adds complexity right now, but it will make it much easier to add
the upcoming OCR feature cleanly. Our goal here is to support "two-tier"
work items, where the first tier might be a PDF to OCR, and the second
tier might be an individual page to send to the LLM.

This will require doing some "fun" async queuing stuff with
backpressure, but in the long run, this model is likely to keep caller
code fairly simple.

We also fix a non-deterministic test.